Moving process factory to custom_python
