+++ applications/ContactStructuralMechanicsApplication/CMakeLists.txt
@@ -73,7 +73,7 @@ set( KRATOS_CONTACT_STRUCTURAL_MECHANICS_APPLICATION_PYTHON_INTERFACE
     ${CMAKE_CURRENT_SOURCE_DIR}/custom_python/add_custom_processes_to_python.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/custom_python/add_custom_linear_solvers_to_python.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/custom_python/add_custom_frictional_laws_to_python.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/custom_utilities/process_factory_utility.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/custom_python/process_factory_utility.cpp
 )
 
 add_library(KratosContactStructuralMechanicsCore SHARED ${KRATOS_CONTACT_STRUCTURAL_MECHANICS_APPLICATION_CORE} ${KRATOS_CONTACT_STRUCTURAL_MECHANICS_APPLICATION_TESTING_SOURCES})
