+++ applications/ContactStructuralMechanicsApplication/custom_python/add_custom_strategies_to_python.cpp
@@ -16,7 +16,7 @@
 // Project includes
 #include "includes/define.h"
 #include "includes/define_python.h"
-#include "custom_utilities/process_factory_utility.h"
+#include "custom_python/process_factory_utility.h"
 #include "custom_python/add_custom_strategies_to_python.h"
 #include "spaces/ublas_space.h"
